+++ package.json
@@ -34,6 +34,7 @@
     "react-dom": "^18.2.0",
     "react-hot-toast": "^2.4.1",
     "react-router-dom": "^6.15.0",
+    "redis-cli": "^2.1.2",
     "stripe": "^13.4.0",
     "tailwindcss": "3.3.3",
     "ts-node": "^10.9.1"
@@ -42,11 +43,11 @@
     "@types/node": "^20.5.8",
     "@types/react": "^18.2.21",
     "@types/react-dom": "^18.2.7",
-    "nanoid": "4.0.2",
-    "redis": "^4.6.8",
-    "typescript": "^5.2.2",
     "eslint-config-prettier": "^9.0.0",
+    "nanoid": "4.0.2",
     "prettier": "^3.0.3",
-    "prettier-plugin-tailwindcss": "^0.5.4"
+    "prettier-plugin-tailwindcss": "^0.5.4",
+    "redis": "^4.6.8",
+    "typescript": "^5.2.2"
   }
 }
